third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/inline/iterable.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap iterable tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#the-stylepropertymap">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<div id="target-empty"></div>
<div id="target-multiple" style="--A: A; width: 10px; --C: C; transition-duration: 1s, 2s; color: red; --B: B;"></div>
<script>
'use strict';

test(() => {
  const styleMap = document.getElementById('target-empty').attributeStyleMap;
  const properties = Array.from(styleMap);
  assert_equals(properties.length, 0);
}, 'Iterating over an empty StylePropertyMap gives a zero-length array');

test(() => {
  const styleMap = document.getElementById('target-multiple').attributeStyleMap;
  const keys = Array.from(styleMap.keys());
  assert_array_equals(keys,
    ['color', 'transition-duration', 'width', '--A', '--B', '--C']);
}, 'StylePropertyMap iterates properties in correct order');

test(() => {
  const styleMap = document.getElementById('target-multiple').attributeStyleMap;
  const [key, result] = [...styleMap][2];
  assert_equals(key, 'width');
  assert_style_value_equals(result, CSS.px(10));
}, 'StylePropertyMap iterator returns CSS properties with the correct CSSStyleValue');

test(() => {
  const styleMap = document.getElementById('target-multiple').attributeStyleMap;
  const [key, result] = [...styleMap][1];
  assert_equals(key, 'transition-duration');
  assert_style_value_array_equals(result, [CSS.s(1), CSS.s(2)]);
}, 'StylePropertyMap iterator returns list-valued properties with the correct CSSStyleValue');

test(() => {
  const styleMap = document.getElementById('target-multiple').attributeStyleMap;
  const entries = [...styleMap];
  assert_equals(entries.length, 6);

  const custom_entries = entries.slice(3);
  for (let i = 0; i < custom_entries.length; i++) {
    const expected_var = ['A', 'B', 'C'][i];
    const [key, result] = custom_entries[i];

    assert_equals(key, '--' + expected_var);
    assert_style_value_equals(result, new CSSUnparsedValue(name));
  }
}, 'StylePropertyMap iterator returns custom properties with the correct CSSStyleValue');

</script>
